Fixed I2C issue! Reset the bus when we start and change the AFIO macro to reg_modify, preserving the debugging settings.
Updated t1000 source to use int rather than uint16.  Fixed
the uvolts to C math with the switch to int and the removal of
the offset from kelvin.

Switch SW D and SW E.  Added a busy count for the I2C bus so we
can reset it if we need to.